+++ VoxelZombieClient/Assets/Scripts/ClientBlockEditor.cs
@@ -19,7 +19,7 @@ namespace Client
             vClient.SendBlockEdit(x, y, z, 0);
         }
 
-        protected override void OnPlaceBlock(ushort x, ushort y, ushort z, ushort blockTag)
+        protected override void OnPlaceBlock(ushort x, ushort y, ushort z, byte blockTag)
         {
             vClient.SendBlockEdit(x, y, z, blockTag);
         }
